requirements.txt
Django
django-extensions
Pillow
MySQLdb-python
djangorestframework